# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /memo/createMemo:
    post:
      operationId: createMemo
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoInputDto"
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoreOutput"
      tags:
        - Memo
  /memo/createCate:
    post:
      operationId: createCate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCateDto"
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCateOutputDto"
      tags:
        - Memo
...
  /memo/updateOneCate:
    post:
      operationId: updateOneCate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOneCateInputDto"
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoreOutput"
      tags:
        - Memo
  /memo/updateManyCate:
    post:
      operationId: updateManyCate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateManyCateInputDto"
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoreOutput"
      tags:
        - Memo
...
info:
  title: My API
  description: An API to do awesome things
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    Promise:
      type: object
      properties: {}
    Tags:
      type: object
      properties:
        id:
          type: number
        createAt:
          format: date-time
          type: string
        updateAt:
          format: date-time
          type: string
        tagName:
          type: string
        user:
          $ref: "#/components/schemas/Promise"
        memos:
          $ref: "#/components/schemas/Promise"
        memoId:
          type: number
        cate:
          $ref: "#/components/schemas/Promise"
        cateId:
          type: number
      required:
        - id
        - createAt
        - updateAt
        - tagName
        - user
        - memos
        - memoId
        - cate
        - cateId
    Memos:
      type: object
      properties:
        id:
          type: number
        createAt:
          format: date-time
          type: string
        updateAt:
          format: date-time
          type: string
        title:
          type: string
        content:
          type: string
        important:
          type: boolean
        user:
          $ref: "#/components/schemas/Promise"
        cate:
          $ref: "#/components/schemas/Promise"
        cateId:
          type: number
        tags:
          $ref: "#/components/schemas/Tags"
      required:
        - id
        - createAt
        - updateAt
        - important
        - user
        - cate
        - cateId
        - tags
    Categories:
      type: object
      properties:
        id:
          type: number
        createAt:
          format: date-time
          type: string
        updateAt:
          format: date-time
          type: string
        cateName:
          type: string
          nullable: true
        user:
          $ref: "#/components/schemas/Promise"
        memos:
          $ref: "#/components/schemas/Memos"
        tags:
          $ref: "#/components/schemas/Tags"
      required:
        - id
        - createAt
        - updateAt
        - cateName
        - user
        - memos
        - tags
...
    CoreOutput:
      type: object
      properties:
        error:
          type: string
        target:
          type: string
        message:
          type: string
        success:
          type: boolean
      required:
        - success
...
    CreateMemoInputDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        important:
          type: boolean
        tags:
          type: string
      required:
        - title
        - content
        - important
    CreateCateDto:
      type: object
      properties:
        cateName:
          type: string
      required:
        - cateName
    CreateCateOutputDto:
      type: object
      properties:
        error:
          type: string
        target:
          type: string
        message:
          type: string
        success:
          type: boolean
        cate:
          $ref: "#/components/schemas/Categories"
      required:
        - success
...
    UpdateOneCateInputDto:
      type: object
      properties:
        cateName:
          type: string
        cateId:
          type: number
      required:
        - cateName
        - cateId
    IUpdateManyCateInput:
      type: object
      properties:
        cateId:
          type: number
        cateName:
          type: string
      required:
        - cateId
        - cateName
    UpdateManyCateInputDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/IUpdateManyCateInput"
      required:
        - data
...